requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
itsdangerous==2.0.1
click==8.0.1
pupdb==0.1.4
python-dotenv==0.19.0
requests==2.31.0
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.24.4
flask-cors==4.0.0
locust==2.15.1 
